stop and move to try metadata
